Implemented integer promotion

+++ examples/fibonacci.c
@@ -1,16 +1,16 @@
-int printf(char*, int, int);
+int printf(char*, short, int);
 
-int fib(int);
+int fib(short);
 
 int main() {
   printf("Fibonacci sequence from 0 to 10:\n", 0, 0);
-  for (int i = 0; i <= 10; ++i) {
-    printf("%d: %d\n", i, fib(i));
+  for (short i = 0; i <= 10; ++i) {
+    printf("%hu: %d\n", i, fib(i));
   }
   return 0;
 }
 
-int fib(int n) {
+int fib(short n) {
   if (n < 2) {
     return n;
   }
